* 调整代码逻辑 system_info.h 清楚警告

+++ pellets/z_win_utils/system_info.h
@@ -357,7 +357,7 @@ namespace WinUtils
                 if (pData && ulLen > 0)
                 {
                     ::memset(pData, 0, ulLen);
-                    for(long lIndex = 0; lIndex < ulLen; ++lIndex)
+                    for(long lIndex = 0; lIndex < (long)ulLen; ++lIndex)
                     {
                         ::SafeArrayGetElement(pSafeArray, &lIndex, pData + lIndex);
                     }
